[Refactor] Merge LakePrimaryIndex into LakePersistentIndex
Resolve the sync conflict in txn_log_applier.cpp: keep only
storage/lake/lake_persistent_index.h. The HEAD side's
storage/lake/lake_primary_index.h is the header this refactor deletes, and
storage/lake/cdc_util.h is CelerData-only and unused in this file.
