+++ be/src/storage/CMakeLists.txt
@@ -228,7 +228,6 @@ set(STORAGE_FILES
     lake/rowset_update_state.cpp
     lake/segment_pk_iterator.cpp
     lake/meta_file.cpp
-    lake/lake_primary_index.cpp
     lake/versioned_tablet.cpp
     lake/vertical_compaction_task.cpp
     lake/cloud_native_index_compaction_task.cpp
